Reset sourcestack when running a new manifest
To be able to use an include in multiple manifests it's necessary to reset the sourcestack when calling function include from a new manifest. $manifest is a global variable, so to be able to use that I renamed 'local manifest' to 'local include_manifest'. Variable cdistaction is now obsolete and replaced by $manifest